05_mvvm_fixes2
